LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-01-expected.txt

CONSOLE MESSAGE: Refused to load script from 'http://localhost:8000/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
